Make the plugin build timeout configurable
Plan to use this for configuration-as-code build.
Not aiming at leaving this there forever, but I think that some plugins
with a lot of tests could need a bit more than one hour.

Or some could even do the contrary to get a faster feedback: lower the
default timeout value to have the build fail faster in case something is
wrong.