Skip tests if changeset content is not relevant
If the changeset does not include files that end with ".java" or files
that end with "pom.xml", skip the tests.

Reduce the CI load while still compiling changes which don't modify Java
files or pom files.

Do not skip tests on first build unless the skip test argument is passed
explicitly.

Do not skip tests on null changeSet.

When a changeset is null, it could either be the first build of the
job or a build that was launched by clicking the run link in the user
interface.  In both those cases, tests should be run.