feat: introduces the attribute 'useContainerAgent' to replace 'useAci'
As per https://issues.jenkins.io/browse/INFRA-2918, the ACI agents are
replaced by Kubernetes agents.

Following this change, the library `buildPlugin` has been updated:
attribute `useAci` is deprecated in favor of the new `useContainerAgent`.

To avoid breaking pipelines, a "soft depreciation" policy is implemented for now:
when attribute `useAci` is defined, then a deprecation warning is printed and
`useContainerAgent` is set the value of `useAci`.

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>