feat: implement `artifact-caching-proxy` in `buildPlugin` with config-file-provider (#502)
This adds an artifactCachingProxyEnabled parameter to the buildPlugin function to allow using one of the artifact caching proxies deployed on each provider (AWS, Azure and DigitalOcean) by specifying a Maven settings file.

It's using the Config File Provider plugin to patch the settings with the basic auth required to access an artifact caching proxy (they all have the same credentials). As a result, the logic here is simpler.

The choice of the proxy provider depends on the ARTIFACT_CACHING_PROXY environment variable defined on every agent, except the Azure VM ones as the corresponding plugin doesn't allow setting env var. (This is also why the default artifact caching proxy provider is set to azure).