Revert "Use Windows VM until Azure ACI outage is resolved (#898)"
Windows virtual machines are configured with Java 11 by default,
though they have Java 17 and Java 21 installed.  This change does not
propagate the Java version configuration to the Windows virtual machine.
Jobs that are configured to use Java 17 on Windows or Java 21 on Windows
are mistakenly using Java 11 on Windows.

Reverting the change so that a complete fix can be tested and confirmed
working before a pull request is proposed.

This reverts commit bcd195f5e1694be916fd883f30361592171d5f47.